# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.CastCreditResponse:
    properties:
      created_at:
        example: "2025-01-01T00:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Jane Doe
        type: string
      picture_url:
        example: https://cdn.example.com/cast/picture.jpg
        type: string
      role:
        example: actor
        type: string
      sort_order:
        example: 0
        type: integer
      updated_at:
        example: "2025-01-01T00:00:00Z"
        type: string
      video_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.CastMemberResponse:
    properties:
      created_at:
        example: "2025-01-01T00:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Jane Doe
        type: string
      picture_url:
        example: https://cdn.example.com/cast/picture.jpg
        type: string
      updated_at:
        example: "2025-01-01T00:00:00Z"
        type: string
    type: object
  github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.CastVideoResponse:
    properties:
      content_rating:
        example: PG
        type: string
      created_at:
        example: "2025-01-01T00:00:00Z"
        type: string
      duration:
        example: 3600
        type: integer
      episode_number:
        example: 3
        type: integer
      role:
        example: actor
        type: string
      season_number:
        example: 1
        type: integer
      series_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      slug:
        example: my-awesome-video-a3f8b2c1
        type: string
      sort_order:
        example: 0
        type: integer
      status:
        example: published
        type: string
      thumbnail_url:
        example: https://cdn.example.com/thumbnails/video.jpg
        type: string
      title:
        example: My Awesome Video
        type: string
      updated_at:
        example: "2025-01-01T00:00:00Z"
        type: string
      video_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.CreateCastRequest:
    properties:
      name:
        example: Jane Doe
        maxLength: 255
        minLength: 1
        type: string
      role:
        example: actor
        maxLength: 100
        minLength: 1
        type: string
      sort_order:
        example: 0
        minimum: 0
        type: integer
    required:
    - name
    - role
    type: object
...
  github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.UpdateCastRequest:
    properties:
      name:
        example: Jane Doe
        maxLength: 255
        minLength: 1
        type: string
    type: object
  github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.UpdateCreditRequest:
    properties:
      role:
        example: director
        maxLength: 100
        minLength: 1
        type: string
      sort_order:
        example: 1
        minimum: 0
        type: integer
    type: object
...
  github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorBody'
    type: object
  github_com_zenfulcode_zencial_internal_pkg_httputil.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_zenfulcode_zencial_internal_pkg_httputil.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.Meta'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@zencial.com
    name: Zencial API Support
  description: Video on Demand streaming platform API
  license:
    name: Proprietary
    url: http://zencial.com/license/
  termsOfService: http://zencial.com/terms/
  title: Zencial VOD API
  version: "1.0"
paths:
...
  /admin/cast:
    get:
      description: Returns a paginated list of all cast members ordered by name. Publisher
        or admin access required.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.CastMemberResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all cast members
      tags:
      - cast
...
  /cast/{id}:
    delete:
      description: Permanently removes a cast member and all their video credits.
        Admin only.
      parameters:
      - description: Cast member ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete cast member globally
      tags:
      - cast
    put:
      consumes:
      - application/json
      description: Updates a cast member's name globally. Admins are unrestricted;
        publishers must have the member credited on one of their videos.
      parameters:
      - description: Cast member ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.UpdateCastRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.CastMemberResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update cast member name
      tags:
      - cast
  /cast/{id}/picture:
    put:
      consumes:
      - multipart/form-data
      description: Uploads or replaces a cast member's picture via multipart form.
      parameters:
      - description: Cast member ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Picture image (JPEG, PNG, WebP, or GIF)
        in: formData
        name: picture
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.CastMemberResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload cast member picture
      tags:
      - cast
  /cast/{id}/videos:
    get:
      description: Returns a paginated list of published videos a cast member appears
        in, ordered by release date (newest first).
      parameters:
      - description: Cast member ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.CastVideoResponse'
                  type: array
                meta:
                  $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.Meta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
      summary: List videos for a cast member
      tags:
      - cast
...
  /videos/{id}/cast:
    get:
      description: Returns all cast credits for a video, ordered by sort_order
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.CastCreditResponse'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
      summary: List cast
      tags:
      - cast
    post:
      consumes:
      - application/json
      description: Adds a cast member to a video using find-or-create by name. Publishers
        may only add cast to their own videos.
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      - description: Cast data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.CreateCastRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.CastCreditResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add cast member to video
      tags:
      - cast
...
  /videos/{videoID}/cast/{castID}:
    delete:
      description: Removes a cast member's credit from a specific video. Publishers
        may only remove credits from their own videos.
      parameters:
      - description: Video ID
        in: path
        name: videoID
        required: true
        type: string
      - description: Cast member ID
        in: path
        name: castID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove cast member from video
      tags:
      - cast
    put:
      consumes:
      - application/json
      description: Updates the role or sort_order for a cast member's credit on a
        specific video. Publishers may only update credits for their own videos.
      parameters:
      - description: Video ID
        in: path
        name: videoID
        required: true
        type: string
      - description: Cast member ID
        in: path
        name: castID
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.UpdateCreditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_zenfulcode_zencial_internal_adapter_handler_v1_dto.CastCreditResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_zenfulcode_zencial_internal_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update cast credit
      tags:
      - cast
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"